+++ CHANGELOG.md
@@ -13,6 +13,26 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Normalize rules input to ensure Laravel 9 support https://github.com/nuwave/lighthouse/pull/2070
 
+## v5.42.0
+
+### Added
+
+- Batched federation entity resolvers https://github.com/nuwave/lighthouse/pull/2043
+
+## v5.41.0
+
+### Added
+
+- Add testing trait `RefreshesSchemaCache` https://github.com/nuwave/lighthouse/pull/2076
+
+### Fixed
+
+- Add proper error handling for invalid schema cache file contents https://github.com/nuwave/lighthouse/pull/2076
+
+### Deprecated
+
+- Deprecate testing trait `ClearsSchemaCache` https://github.com/nuwave/lighthouse/pull/2076
+
 ## v5.40.1
 
 ### Fixed